requirements.txt
numpy~=1.19.2
pillow~=8.4.0
torchvision~=0.7.0
scikit-image~=0.18.3
tensorflow~=2.1.0